+++ package.json
@@ -25,7 +25,6 @@
     "cors": "^2.8.5",
     "dotenv": "^6.2.0",
     "express": "^4.16.2",
-    "jest": "^22.3.0",
     "knex": "^0.16.3",
     "morgan": "^1.9.1",
     "mysql": "^2.16.0",
@@ -37,6 +36,7 @@
   "devDependencies": {
     "css-loader": "^2.0.2",
     "faker": "^4.1.0",
+    "jest": "^22.4.4",
     "nodemon": "^1.18.9",
     "pm2": "^3.2.4",
     "style-loader": "^0.23.1",
