--- sdc/fake-data/projectNameParts.js
@@ -1,1007 +1,0 @@
-module.exports = {
-  projectNouns: [
-    'soup',
-    'market',
-    'reading',
-    'snake',
-    'coal',
-    'toothbrush',
-    'meeting',
-    'lip',
-    'border',
-    'insurance',
-    'low',
-    'route',
-    'furniture',
-    'back',
-    'bat',
-    'sky',
-    'bear',
-    'tray',
-    'hat',
-    'skirt',
-    'vase',
-    'mind',
-    'bite',
-    'day',
-    'theory',
-    'minister',
-    'tooth',
-    'kitty',
-    'hour',
-    'statement',
-    'face',
-    'sock',
-    'books',
-    'sweater',
-    'expansion',
-    'door',
-    'loaf',
-    'potato',
-    'mice',
-    'temper',
-    'locket',
-    'turkey',
-    'bells',
-    'behavior',
-    'believe',
-    'toy',
-    'friction',
-    'mother',
-    'self',
-    'play',
-    'arm',
-    'needle',
-    'mine',
-    'bomb',
-    'badge',
-    'partner',
-    'fish',
-    'education',
-    'regret',
-    'dust',
-    'zebra',
-    'quartz',
-    'angle',
-    'lock',
-    'square',
-    'condition',
-    'van',
-    'finger',
-    'dog',
-    'prose',
-    'alarm',
-    'anger',
-    'bait',
-    'heat',
-    'volcano',
-    'spot',
-    'card',
-    'wilderness',
-    'poison',
-    'mint',
-    'ball',
-    'paper',
-    'frogs',
-    'cherries',
-    'cable',
-    'touch',
-    'fork',
-    'jar',
-    'distribution',
-    'look',
-    'morning',
-    'discussion',
-    'pull',
-    'milk',
-    'cook',
-    'can',
-    'flavor',
-    'sheep',
-    'dirt',
-    'mailbox',
-    'smell',
-    'support',
-    'dolls',
-    'frog',
-    'horses',
-    'monkey',
-    'carriage',
-    'road',
-    'nest',
-    'teeth',
-    'song',
-    'cracker',
-    'quarter',
-    'rainstorm',
-    'north',
-    'voice',
-    'clam',
-    'part',
-    'afterthought',
-    'town',
-    'pig',
-    'idea',
-    'rhythm',
-    'finger',
-    'cave',
-    'interest',
-    'stem',
-    'tail',
-    'vessel',
-    'unit',
-    'cobweb',
-    'mask',
-    'calendar',
-    'whistle',
-    'skate',
-    'weight',
-    'table',
-    'creature',
-    'nation',
-    'leather',
-    'fang',
-    'flag',
-    'amount',
-    'spy',
-    'branch',
-    'animal',
-    'impulse',
-    'observation',
-    'snails',
-    'caption',
-    'planes',
-    'degree',
-    'store',
-    'toes',
-    'hole',
-    'taste',
-    'spade',
-    'detail',
-    'jeans',
-    'duck',
-    'stream',
-    'beef',
-    'wren',
-    'machine',
-    'plants',
-    'home',
-    'cent',
-    'blow',
-    'arithmetic',
-    'sponge',
-    'system',
-    'attraction',
-    'earthquake',
-    'picture',
-    'pizzas',
-    'gold',
-    'cow',
-    'trade',
-    'selection',
-    'crib',
-    'account',
-    'title',
-    'school',
-    'jelly',
-    'fire',
-    'base',
-    'shoes',
-    'shirt',
-    'camera',
-    'rake',
-    'steam',
-    'waves',
-    'trick',
-    'butter',
-    'motion',
-    'ground',
-    'agreement',
-    'quince',
-    'station',
-    'harmony',
-    'summer',
-    'scarecrow',
-    'wrench',
-    'silver',
-    'tongue',
-    'laugh',
-    'geese',
-    'flowers',
-    'scale',
-    'stop',
-    'respect',
-    'loss',
-    'cakes',
-    'fog',
-    'passenger',
-    'fly',
-    'wheel',
-    'move',
-    'icicle',
-    'run',
-    'chicken',
-    'hand',
-    'yak',
-    'death',
-    'friend',
-    'swim',
-    'powder',
-    'wish',
-    'shock',
-    'snow',
-    'head',
-    'chalk',
-    'error',
-    'form',
-    'advice',
-    'texture',
-    'breath',
-    'toys',
-    'pest',
-    'camp',
-    'wool',
-    'shelf',
-    'ship',
-    'society',
-    'destruction',
-    'income',
-    'rice',
-    'thought',
-    'action',
-    'tomatoes',
-    'rat',
-    'rain',
-    'straw',
-    'chin',
-    'scent',
-    'wax',
-    'legs',
-    'steel',
-    'water',
-    'tiger',
-    'slave',
-    'walk',
-    'team',
-    'queen',
-    'authority',
-    'connection',
-    'waste',
-    'linen',
-    'uncle',
-    'apparel',
-    'basket',
-    'hose',
-    'visitor',
-    'division',
-    'balance',
-    'cherry',
-    'knowledge',
-    'achiever',
-    'button',
-    'sense',
-    'range',
-    'girls',
-    'flight',
-    'rabbit',
-    'instrument',
-    'mitten',
-    'book',
-    'weather',
-    'experience',
-    'writing',
-    'spark',
-    'flesh',
-    'show',
-    'horse',
-    'trains',
-    'level',
-    'oranges',
-    'plot',
-    'business',
-    'things',
-    'ocean',
-    'whip',
-    'hall',
-    'relation',
-    'digestion',
-    'produce',
-    'teaching',
-    'start',
-    'pies',
-    'fuel',
-    'magic',
-    'meal',
-    'grandmother',
-    'car',
-    'dress',
-    'orange',
-    'bird',
-    'feeling',
-    'voyage',
-    'airport',
-    'drawer',
-    'rose',
-    'trail',
-    'bushes',
-    'twist',
-    'thrill',
-    'grandfather',
-    'quiet',
-    'current',
-    'stew',
-    'salt',
-    'building',
-    'sister',
-    'fall',
-    'pin',
-    'ducks',
-    'time',
-    'cause',
-    'comparison',
-    'sand',
-    'smoke',
-    'sugar',
-    'flame',
-    'office',
-    'plane',
-    'birds',
-    'elbow',
-    'shape',
-    'quill',
-    'year',
-    'floor',
-    'blood',
-    'chickens',
-    'committee',
-    'arch',
-    'seed',
-    'airplane',
-    'advertisement',
-    'land',
-    'trousers',
-    'zinc',
-    'point',
-    'mom',
-    'tree',
-    'minute',
-    'spoon',
-    'crow',
-    'step',
-    'view',
-    'pickle',
-    'rabbits',
-    'pet',
-    'snail',
-    'liquid',
-    'group',
-    'trees',
-    'stranger',
-    'earth',
-    'doctor',
-    'health',
-    'wave',
-    'window',
-    'turn',
-    'pan',
-    'ice',
-    'smash',
-    'history',
-    'bone',
-    'sun',
-    'match',
-    'throne',
-    'laborer',
-    'basketball',
-    'foot',
-    'deer',
-    'collar',
-    'meat',
-    'month',
-    'brick',
-    'moon',
-    'bubble',
-    'quicksand',
-    'pleasure',
-    'thunder',
-    'truck',
-    'yard',
-    'dogs',
-    'receipt',
-    'baseball',
-    'vest',
-    'knee',
-    'crime',
-    'peace',
-    'toe',
-    'hill',
-    'position',
-    'curtain',
-    'knife',
-    'top',
-    'reason',
-    'cat',
-    'umbrella',
-    'lace',
-    'wash',
-    'drain',
-    'copper',
-    'zephyr',
-    'lunch',
-    'kettle',
-    'reward',
-    'noise',
-    'shade',
-    'friends',
-    'man',
-    'downtown',
-    'dinosaurs',
-    'hair',
-    'river',
-    'end',
-    'trouble',
-    'stocking',
-    'bit',
-    'jump',
-    'root',
-    'channel',
-    'sleep',
-    'fact',
-    'price',
-    'swing',
-    'plantation',
-    'street',
-    'size',
-    'letter',
-    'increase',
-    'religion',
-    'celery',
-    'name',
-    'porter',
-    'stick',
-    'person',
-    'veil',
-    'star',
-    'bridge',
-    'talk',
-    'ladybug',
-    'rate',
-    'stomach',
-    'bell',
-    'dad',
-    'ghost',
-    'hook',
-    'industry',
-    'tank',
-    'toothpaste',
-    'humor',
-    'slip',
-    'stone',
-    'sound',
-    'hospital',
-    'dinner',
-    'berry',
-    'snakes',
-    'pipe',
-    'playground',
-    'neck',
-    'chance',
-    'test',
-    'boundary',
-    'knot',
-    'clover',
-    'oatmeal',
-    'canvas',
-    'pump',
-    'birthday',
-    'cows',
-    'jail',
-    'horn',
-    'woman',
-    'lettuce',
-    'soda'
-  ],
-
-  verbs: [
-    'arrange',
-    'work',
-    'whip',
-    'exist',
-    'add',
-    'seal',
-    'kill',
-    'dam',
-    'satisfy',
-    'realise',
-    'number',
-    'label',
-    'knot',
-    'expand',
-    'tempt',
-    'permit',
-    'squeal',
-    'print',
-    'file',
-    'tremble',
-    'depend',
-    'close',
-    'produce',
-    'pump',
-    'remember',
-    'afford',
-    'retire',
-    'blush',
-    'decay',
-    'bump',
-    'clear',
-    'complain',
-    'spray',
-    'smash',
-    'mess up',
-    'challenge',
-    'rule',
-    'joke',
-    'tug',
-    'jog',
-    'signal',
-    'communicate',
-    'comb',
-    'wonder',
-    'settle',
-    'sign',
-    'wink',
-    'polish',
-    'wreck',
-    'instruct',
-    'wish',
-    'balance',
-    'snow',
-    'wave',
-    'attack',
-    'program',
-    'reach',
-    'store',
-    'post',
-    'step',
-    'disapprove',
-    'shop',
-    'suck',
-    'open',
-    'tick',
-    'thaw',
-    'plant',
-    'introduce',
-    'harm',
-    'note',
-    'cycle',
-    'injure',
-    'irritate',
-    'claim',
-    'protect',
-    'fool',
-    'drag',
-    'tow',
-    'punish',
-    'obey',
-    'coil',
-    'peel',
-    'list',
-    'collect',
-    'exercise',
-    'battle',
-    'ski',
-    'develop',
-    'suspend',
-    'blot',
-    'cure',
-    'confuse',
-    'fade',
-    'nod',
-    'heat',
-    'whisper',
-    'pull',
-    'suffer',
-    'jail',
-    'interest',
-    'hover',
-    'greet',
-    'colour',
-    'repeat',
-    'guard',
-    'boast',
-    'discover',
-    'worry',
-    'attach',
-    'dare',
-    'time',
-    'book',
-    'consider',
-    'moan',
-    'raise',
-    'groan',
-    'meddle',
-    'name',
-    'fax',
-    'squeeze',
-    'embarrass',
-    'rejoice',
-    'invent',
-    'overflow',
-    'appear',
-    'pour',
-    'save',
-    'bruise',
-    'return',
-    'skip',
-    'entertain',
-    'rush',
-    'bounce',
-    'repair',
-    'fail',
-    'grip',
-    'consist',
-    'beam',
-    'bore',
-    'radiate',
-    'spoil',
-    'reply',
-    'yell',
-    'preserve',
-    'screw',
-    'surprise',
-    'bow',
-    'transport',
-    'belong',
-    'attend',
-    'strengthen',
-    'curl',
-    'like',
-    'float',
-    'test',
-    'excite',
-    'tame',
-    'tour',
-    'fire',
-    'yawn',
-    'zoom',
-    'call',
-    'sip',
-    'escape',
-    'burn',
-    'record',
-    'jam',
-    'weigh',
-    'license',
-    'nest',
-    'match',
-    'decide',
-    'fix',
-    'hum',
-    'damage',
-    'heap',
-    'touch',
-    'agree',
-    'rely',
-    'fancy',
-    'enjoy',
-    'scatter',
-    'amuse',
-    'excuse',
-    'prefer',
-    'knock',
-    'employ',
-    'pedal',
-    'tire',
-    'terrify',
-    'shiver',
-    'queue',
-    'mend',
-    'sprout',
-    'stain',
-    'cough',
-    'search',
-    'plan',
-    'shave',
-    'boil',
-    'dress',
-    'provide',
-    'join',
-    'pick',
-    'bathe',
-    'spill',
-    'grate',
-    'follow',
-    'fear',
-    'admit',
-    'whistle',
-    'listen',
-    'sail',
-    'suit',
-    'matter',
-    'blind',
-    'hug',
-    'pop',
-    'answer',
-    'coach',
-    'command',
-    'kneel',
-    'frame',
-    'miss',
-    'squeak',
-    'hope',
-    'increase',
-    'order',
-    'contain',
-    'spark',
-    'accept',
-    'prepare',
-    'ignore',
-    'care',
-    'start',
-    'mine',
-    'earn',
-    'soothe',
-    'sniff',
-    'glow',
-    'chop',
-    'avoid',
-    'haunt',
-    'fasten',
-    'phone',
-    'scratch',
-    'admire',
-    'shrug',
-    'squash',
-    'judge',
-    'head',
-    'pause',
-    'prevent',
-    'please',
-    'peep',
-    'stare',
-    'fetch',
-    'fit',
-    'release',
-    'march',
-    'cover',
-    'live',
-    'waste',
-    'suspect',
-    'concentrate',
-    'supply',
-    'long',
-    'bake',
-    'check',
-    'bless',
-    'detect',
-    'concern',
-    'deserve',
-    'identify',
-    'obtain',
-    'jump',
-    'stuff',
-    'occur',
-    'warn',
-    'appreciate',
-    'pack',
-    'pine',
-    'change',
-    'possess',
-    'box',
-    'snatch',
-    'rob',
-    'charge',
-    'wait',
-    'disarm',
-    'murder',
-    'push',
-    'hunt',
-    'help',
-    'compete',
-    'attempt',
-    'analyse',
-    'reflect',
-    'warm',
-    'back',
-    'smile',
-    'multiply',
-    'invite',
-    'shelter',
-    'influence',
-    'empty',
-    'rock',
-    'inform',
-    'apologise',
-    'applaud',
-    'switch',
-    'rhyme',
-    'talk',
-    'mark',
-    'drip',
-    'stroke',
-    'nail',
-    'curve',
-    'remain',
-    'trouble',
-    'steer',
-    'announce',
-    'end',
-    'stretch',
-    'unpack',
-    'force',
-    'rinse',
-    'camp',
-    'tickle',
-    'water',
-    'wriggle',
-    'offer',
-    'support',
-    'promise',
-    'lick',
-    'level',
-    'delight',
-    'crawl',
-    'flower',
-    'dislike',
-    'mug',
-    'pinch',
-    'rot',
-    'handle',
-    'drown',
-    'prick',
-    'interfere',
-    'copy',
-    'hurry',
-    'taste',
-    'hook',
-    'decorate',
-    'bubble',
-    'buzz',
-    'brake',
-    'interrupt',
-    'fry',
-    'heal',
-    'paste',
-    'trip',
-    'extend',
-    'annoy',
-    'stitch',
-    'explain',
-    'film',
-    'manage',
-    'clip',
-    'blink',
-    'move',
-    'need',
-    'zip',
-    'paddle',
-    'imagine',
-    'stay',
-    'flow',
-    'complete',
-    'crack',
-    'drum',
-    'snore',
-    'hop',
-    'rescue',
-    'sparkle',
-    'thank',
-    'tease',
-    'tap',
-    'stop',
-    'shade',
-    'smoke',
-    'impress',
-    'lighten',
-    'lock',
-    'love',
-    'fence',
-    'share',
-    'replace',
-    'muddle',
-    'reduce',
-    'trick',
-    'place',
-    'scribble',
-    'recognise',
-    'gaze',
-    'breathe',
-    'happen',
-    'measure',
-    'guarantee',
-    'preach',
-    'delay',
-    'use',
-    'approve',
-    'shock',
-    'report',
-    'connect',
-    'correct',
-    'bang',
-    'melt',
-    'examine',
-    'train',
-    'beg',
-    'saw',
-    'enter',
-    'paint',
-    'puncture',
-    'trot',
-    'sigh',
-    'wipe',
-    'improve',
-    'hammer',
-    'spot',
-    'form',
-    'bare',
-    'trap',
-    'hate',
-    'x-ray',
-    'stamp',
-    'vanish',
-    'travel',
-    'describe',
-    'object',
-    'cross',
-    'scare',
-    'mourn',
-    'gather',
-    'unite',
-    'behave',
-    'dance',
-    'tie',
-    'sack',
-    'memorise',
-    'knit',
-    'perform',
-    'doubt',
-    'spare',
-    'double',
-    'slow',
-    'strap',
-    'chase',
-    'grease',
-    'trace',
-    'land',
-    'laugh',
-    'scold',
-    'flap',
-    'separate',
-    'watch',
-    'tumble',
-    'expect',
-    'rub',
-    'pat',
-    'try',
-    'whine',
-    'regret',
-    'fold',
-    'arrest',
-    'mix',
-    'slip',
-    'trust',
-    'telephone',
-    'bury',
-    'remind',
-    'argue',
-    'slap',
-    'man',
-    'stir',
-    'continue',
-    'surround',
-    'scorch',
-    'welcome',
-    'found',
-    'drain',
-    'hang',
-    'undress',
-    'subtract',
-    'guide',
-    'clap',
-    'turn',
-    'smell',
-    'owe',
-    'press',
-    'poke'
-  ]
-}
